[refactor] flatten exp package
Flatten cmd/exp into cmd, prefixing each file with exp_. This should
make it easier to DRY up some code in these two packages.
