+++ cmd/exp.go
@@ -1,9 +1,13 @@
-package exp
+package cmd
 
 import "github.com/spf13/cobra"
 
+func init() {
+	rootCmd.AddCommand(expCmd)
+}
+
 // ExpCmd is a subcommand for experimental commands
-var ExpCmd = &cobra.Command{
+var expCmd = &cobra.Command{
 	Use:          "exp",
 	Short:        "Experimental commands",
 	Long:         "Grouping of experimental commands. These are experimental and prone to change",
