+++ cmd/exp_auto_remote_state.go
@@ -1,4 +1,4 @@
-package exp
+package cmd
 
 import (
 	"os"
@@ -13,7 +13,7 @@ func init() {
 	autoRemoteStateCmd.Flags().String("path", "", "path to a working directory")
 	autoRemoteStateCmd.Flags().StringP("config", "c", "fogg.yml", "Use this to override the fogg config file.")
 
-	ExpCmd.AddCommand(autoRemoteStateCmd)
+	expCmd.AddCommand(autoRemoteStateCmd)
 }
 
 var autoRemoteStateCmd = &cobra.Command{
