+++ cmd/exp_examine.go
@@ -1,17 +1,16 @@
-package exp
+package cmd
 
 import (
 	"os"
 
 	"github.com/chanzuckerberg/fogg/errs"
 	"github.com/chanzuckerberg/fogg/exp/examine"
-	"github.com/sirupsen/logrus"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 )
 
 func init() {
-	ExpCmd.AddCommand(examineCmd)
+	expCmd.AddCommand(examineCmd)
 }
 
 //TODO:(EC) Create a flag for path to walk
@@ -32,13 +31,4 @@ var examineCmd = &cobra.Command{
 
 		return examine.Examine(fs, pwd)
 	},
-}
-
-func openGitOrExit(fs afero.Fs) {
-	_, err := fs.Stat(".git")
-	if err != nil {
-		// assuming this means no repository
-		logrus.Fatal("fogg must be run from the root of a git repo")
-		os.Exit(1)
-	}
 }
